requirements.txt
streamlit==1.32.0
gspread==6.0.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.125.0
Pillow==10.3.0
pandas==2.2.1
streamlit-calendar==1.0.0
